@@ -10,13 +10,15 @@
     "chart.js": "^2.7.2",
     "immutable": "^3.8.1",
     "jquery": "^3.0.0",
-    "lodash": "^4.17.4",
+    "js-beautify": "^1.7.5",
+    "lodash": "^4.17.10",
     "match-sorter": "^2.2.1",
     "material-ui": "^1.0.0-beta.29",
     "material-ui-icons": "^1.0.0-beta.17",
     "moment-timezone": "^0.5.16",
     "pg": "^7.4.1",
-    "react": "^16.2.0",
+    "react": "^16.4.1",
+    "react-copy-to-clipboard": "^5.0.1",
     "react-d3-graph": "^1.0.1",
     "react-dom": "^16.2.0",
     "react-event-timeline": "^1.5.1",
@@ -26,7 +28,7 @@
     "react-refetch": "^1.0.4",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
-    "react-scripts": "1.1.0",
+    "react-scripts": "1.1.4",
     "react-select": "^1.2.1",
     "react-slick": "^0.18.1",
     "react-swipeable-views": "^0.12.12",
@@ -34,7 +36,7 @@
     "react-timeago": "^4.1.9",
     "react-tree-graph": "^3.1.0",
     "react-websocket": "^2.0.0",
-    "reactstrap": "^5.0.0-beta",
+    "reactstrap": "^5.0.0",
     "recharts": "^1.0.0-beta.9",
     "redux": "^3.7.2",
     "redux-actions": "^0.9.1",
@@ -45,15 +47,15 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
+    "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
   },
   "proxy": "http://localhost:8080/",
   "devDependencies": {
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
-    "jest": "^22.2.1",
     "jest-enzyme": "^4.2.0",
+    "react-addons-test-utils": "^15.6.2",
     "react-test-renderer": "^16.2.0",
     "redux-devtools": "^3.4.1",
     "redux-mock-store": "^1.5.1",
