@@ -13,32 +13,65 @@
  See the License for the specific language governing permissions and
  limitations under the License.
  */
-'use strict';
-var log4js = require('log4js');
-var logger = log4js.getLogger('Helper');
-logger.setLevel('INFO');
 
-var path = require('path');
-var fs = require('fs-extra');
+"use strict";
+var log4js = require("log4js/lib/log4js");
+var appList = [];
+
+var path = require("path");
+var fs = require("fs-extra");
 
 function readAllFiles(dir) {
-	var files = fs.readdirSync(dir);
-	var certs = [];
-	files.forEach((file_name) => {
-		let file_path = path.join(dir, file_name);
-		let data = fs.readFileSync(file_path);
-		certs.push(data);
-	});
-	return certs;
+  var files = fs.readdirSync(dir);
+  var certs = [];
+  files.forEach(file_name => {
+    let file_path = path.join(dir, file_name);
+    let data = fs.readFileSync(file_path);
+    certs.push(data);
+  });
+  return certs;
 }
 
+/*
+Please assign the logger with the filename for the application logging
+and assign the logger with "pgservice" for database logging for any filename. Please find an example below.
+To stacktrace, please pass the error.stack object to the logger. If there is no error.stack object pass in a
+string with description.
 
-var getLogger = function (moduleName) {
-	var logger = log4js.getLogger(moduleName);
-	logger.setLevel('INFO');
-	return logger;
+var helper = require("./app/helper");
+var logger = helper.getLogger("main");
+logger.setLevel('INFO');
+
+
+*/
+
+var getLogger = function(moduleName) {
+  if (moduleName == "pgservice") {
+    var logger = log4js.getLogger("pgservice");
+  } else {
+    appList.push(moduleName);
+    var logger = log4js.getLogger(moduleName);
+  }
+
+  log4js.configure({
+    appenders: [
+      {
+        type: "dateFile",
+        filename: "logs/app/app.log",
+        pattern: "-yyyy-MM-dd",
+        category: appList
+      },
+      {
+        type: "dateFile",
+        filename: "logs/db/db.log",
+        pattern: "-yyyy-MM-dd",
+        category: ["pgservice"]
+      }
+    ]
+  });
+  logger.setLevel("INFO");
+  return logger;
 };
-
 
 exports.getLogger = getLogger;
 exports.readAllFiles = readAllFiles;
