@@ -12,6 +12,7 @@ import lightBlue from 'material-ui/colors/lightBlue';
 import red from 'material-ui/colors/red';
 import HeaderView from '../Header/HeaderView';
 import LandingPage from '../View/LandingPage';
+
 const muiTheme = createMuiTheme({
   palette: {
     contrastThreshold: 3,
@@ -24,28 +25,27 @@ const muiTheme = createMuiTheme({
   },
 });
 
-
 class App extends Component {
   constructor(props) {
     super(props);
     this.state = { loading: true }
   }
+
   componentWillMount() {
     setTimeout(() => this.setState({ loading: false }), 6000);
   }
-  componentDidMount() {
 
-  }
   render() {
     if (this.state.loading) {
       return (
         <LandingPage />
       );
     }
+
     return (
       <MuiThemeProvider theme={muiTheme} >
         <div>
-          <HeaderView />
+            <HeaderView />
           <Main />
         </div>
       </MuiThemeProvider>
