@@ -2,23 +2,119 @@
  *    SPDX-License-Identifier: Apache-2.0
  */
 
-import React from 'react';
-import { Table } from 'reactstrap';
-const Channels = ({ channelList }) => {
+import React, { Component } from 'react';
+import ReactTable from 'react-table';
+import 'react-table/react-table.css';
+import matchSorter from 'match-sorter';
+
+class Channels extends Component {
+  constructor(props) {
+    super(props);
+  }
+
+  componentWillMount() {
+   this.props.getChannels()
+  }
+
+  componentDidMount() {
+    setInterval(() => {
+    this.props.getChannels()
+    }, 600000);
+  }
+
+  reactTableSetup = () => {
+    return [
+      {
+        Header: 'ID',
+        accessor: 'id',
+        filterMethod: (filter, rows) =>
+          matchSorter(
+            rows,
+            filter.value,
+            { keys: ['id'] },
+            { threshold: matchSorter.rankings.SIMPLEMATCH }
+          ),
+        filterAll: true,
+        width: 100
+      },
+      {
+        Header: 'Channel Name',
+        accessor: 'channelname',
+        filterMethod: (filter, rows) =>
+          matchSorter(
+            rows,
+            filter.value,
+            { keys: ['channelname'] },
+            { threshold: matchSorter.rankings.SIMPLEMATCH }
+          ),
+        filterAll: true
+      },
+      {
+        Header: 'Channel Hash',
+        accessor: 'channel_hash',
+        filterMethod: (filter, rows) =>
+          matchSorter(
+            rows,
+            filter.value,
+            { keys: ['channel_hash'] },
+            { threshold: matchSorter.rankings.SIMPLEMATCH }
+          ),
+        filterAll: true
+      },
+      {
+        Header: 'Blocks',
+        accessor: 'blocks',
+        filterMethod: (filter, rows) =>
+          matchSorter(
+            rows,
+            filter.value,
+            { keys: ['blocks'] },
+            { threshold: matchSorter.rankings.SIMPLEMATCH }
+          ),
+        filterAll: true,
+        width: 125
+      }, {
+        Header: 'Transactions',
+        accessor: 'transactions',
+        filterMethod: (filter, rows) =>
+          matchSorter(
+            rows,
+            filter.value,
+            { keys: ['transactions'] },
+            { threshold: matchSorter.rankings.SIMPLEMATCH }
+          ),
+        filterAll: true,
+        width: 125
+      },
+      {
+        Header: 'Timestamp',
+        accessor: 'createdat',
+        filterMethod: (filter, rows) =>
+          matchSorter(
+            rows,
+            filter.value,
+            { keys: ['createdat'] },
+            { threshold: matchSorter.rankings.SIMPLEMATCH }
+          ),
+        filterAll: true
+      }
+    ];
+  };
+
+  render() {
     return (
-        <div>
-            <Table>
-                <thead>
-                    <th>Channel Id</th>
-                </thead>
-                <tbody>
-                    {channelList.map(channel=>
-                    <tr key={channel.channel_id}>
-                        <td>{channel.channel_id} </td>
-                    </tr>)}
-                </tbody>
-            </Table>
-        </div>
+      <div className="blockPage">
+                <ReactTable
+                  data={this.props.channels}
+                  columns={this.reactTableSetup()}
+                  defaultPageSize={5}
+                  className="-striped -highlight"
+                  filterable
+                  minRows={0}
+                />
+      </div>
     );
-};
+  }
+}
+
 export default Channels;
