@@ -9,43 +9,35 @@ import 'react-table/react-table.css';
 import matchSorter from 'match-sorter';
 
 const Peers = ({ peerList }) => {
-
     const columnHeaders = [
         {
-          Header: "Peer Name",
-          accessor: "server_hostname",
-          filterMethod: (filter, rows) =>
-            matchSorter(rows, filter.value, { keys: ["server_hostname"] }, { threshold: matchSorter.rankings.SIMPLEMATCH }),
-          filterAll: true
+            Header: "Peer Name",
+            accessor: "server_hostname",
+            filterMethod: (filter, rows) =>
+                matchSorter(rows, filter.value, { keys: ["server_hostname"] }, { threshold: matchSorter.rankings.SIMPLEMATCH }),
+            filterAll: true
         },
         {
-          Header: "Request Url",
-          accessor: "requests",
-          filterMethod: (filter, rows) =>
-          matchSorter(rows, filter.value, { keys: ["requests"] }, { threshold: matchSorter.rankings.SIMPLEMATCH }),
-        filterAll: true
+            Header: "Request Url",
+            accessor: "requests",
+            filterMethod: (filter, rows) =>
+                matchSorter(rows, filter.value, { keys: ["requests"] }, { threshold: matchSorter.rankings.SIMPLEMATCH }),
+            filterAll: true
         }
-      ];
+    ];
 
     return (
         <div>
-            <Container>
-                <Row>
-                    <Col>
-                        <div className="scrollTable">
-                         <ReactTable
-                            data={peerList}
-                            columns={columnHeaders}
-                            defaultPageSize={5}
-                            className="-striped -highlight"
-                            filterable
-                        />
-                        </div>
-                    </Col>
-                </Row>
-            </Container>
-
+            <ReactTable
+                data={peerList}
+                columns={columnHeaders}
+                defaultPageSize={5}
+                className="-striped -highlight"
+                filterable
+                minRows={0}
+            />
         </div>
     );
 };
+
 export default Peers;
