@@ -2,60 +2,67 @@
  *    SPDX-License-Identifier: Apache-2.0
  */
 
-import React from 'react';
+import React, { Component } from 'react';
+import compose from 'recompose/compose';
 import { withStyles } from 'material-ui/styles';
 import PropTypes from 'prop-types';
-import Card, { CardContent } from 'material-ui/Card';
-import Typography from 'material-ui/Typography';
+import Chaincodes from '../Lists/Chaincodes';
 
 const styles = theme => ({
-    root: {
-        flexGrow: 1,
-        paddingTop: 42,
-        position: 'relative',
-    },
-    card: {
-        height: 250,
-        minWidth: 1290,
-        margin: 20,
-        textAlign: 'left',
-        display: 'inline-block',
-    },
-    title: {
-        fontSize: 16,
-        color: theme.palette.text.secondary,
-        position: 'absolute',
-        left: 40,
-        top: 60
-    },
-    content: {
-        fontSize: 12,
-        color: theme.palette.text.secondary,
-        position: 'absolute',
-        left: 40,
-        top: 70
-    }
+  root: {
+    flexGrow: 1,
+    paddingTop: 42,
+    position: 'relative'
+  },
+  card: {
+    height: 250,
+    minWidth: 1290,
+    margin: 20,
+    textAlign: 'left',
+    display: 'inline-block'
+  },
+  title: {
+    fontSize: 16,
+    color: theme.palette.text.secondary,
+    position: 'absolute',
+    left: 40,
+    top: 60
+  },
+  content: {
+    fontSize: 12,
+    color: theme.palette.text.secondary,
+    position: 'absolute',
+    left: 40,
+    top: 70
+  }
 });
 
-function ChaincodeView(props) {
-    const { classes } = props;
+export class ChaincodeView extends Component {
+  constructor(props) {
+    super(props);
+  }
+
+  render() {
+    const { classes } = this.props;
     return (
-        <div>
-            <Card className={classes.card} title={'Chaincode List'}>
-                <CardContent>
-                    <Typography className={classes.title}>Chaincode List </Typography>
-                </CardContent>
-                <CardContent className={classes.content}>
-                    Chain code details goes here
-                </CardContent>
-            </Card>
+      <div className="view-fullwidth">
+        <div className="view-display">
+          <Chaincodes
+            channel={this.props.channel}
+            countHeader={this.props.countHeader}
+            chaincodes={this.props.chaincodes}
+            getChaincodes={this.props.getChaincodes}
+          />
         </div>
+      </div>
     );
+  }
 }
 
-
 ChaincodeView.propTypes = {
-    classes: PropTypes.object.isRequired,
+  classes: PropTypes.object.isRequired
 };
 
-export default withStyles(styles)(ChaincodeView);
+export default compose(
+  withStyles(styles)
+)(ChaincodeView);
