@@ -2,128 +2,216 @@
  *    SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { Component } from 'react';
-import { withStyles } from 'material-ui/styles';
-import PropTypes from 'prop-types';
-import FontAwesome from 'react-fontawesome';
-import Typography from 'material-ui/Typography';
+import React, { Component } from "react";
+import { withStyles } from "material-ui/styles";
+import PropTypes from "prop-types";
+import FontAwesome from "react-fontawesome";
+import { CopyToClipboard } from "react-copy-to-clipboard";
+import Typography from "material-ui/Typography";
+
+import moment from "moment-timezone";
 import {
-    DialogContent,
-    DialogContentText,
-    DialogTitle,
-} from 'material-ui/Dialog';
-import moment from 'moment-timezone';
+  Table,
+  Card,
+  CardText,
+  CardBody,
+  CardTitle,
+  CardSubtitle,
+  Button
+} from "reactstrap";
 
 const styles = theme => ({
-    root: {
-        flexGrow: 1,
-        paddingTop: 42,
-        position: 'relative',
-    },
-    card: {
-        height: 250,
-        width: 1215,
-        margin: 20,
-        textAlign: 'left',
-        display: 'inline-block',
-    },
-    title: {
-        fontSize: 16,
-        color: theme.palette.text.secondary,
-        position: 'absolute',
-        left: 40,
-        top: 60
-    },
-    content: {
-        fontSize: 12,
-        color: theme.palette.text.secondary,
-        position: 'absolute',
-        left: 40,
-        top: 70
-    }
+  root: {
+    flexGrow: 1,
+    paddingTop: 42,
+    position: "relative"
+  }
 });
+const reads = {
+  color: "#2AA233"
+};
+const writes = {
+  color: "#DD8016"
+};
 
-class TransactionView extends Component {
-    constructor(props, context) {
-        super(props, context);
-        this.state = {
-            loading: false
-        }
-    }
-    componentWillReceiveProps(nextProps) {
-        this.setState({ loading: false });
-    }
-    render() {
-        const { classes } = this.props;
-        if (this.props.transaction.read_set === undefined) {
-            return (
-                <div>
-                    <DialogTitle>Transaction Detail</DialogTitle>
-                    <DialogContent>
-                        <DialogContentText>
-                            <p className="loading-wheel"> <FontAwesome name="circle-o-notch" size="3x" spin /></p>
-                        </DialogContentText>
-                    </DialogContent>
-                </div>
-            );
-        }
-        else {
-            return (
-                <div>
-                    <DialogTitle>Transaction Detail</DialogTitle>
-                    <DialogContent>
-                        <DialogContentText>
+export class TransactionView extends Component {
+  constructor(props, context) {
+    super(props, context);
+    this.state = {
+      loading: false
+    };
+  }
 
-                            <b>Tx:</b>{this.props.transaction.txhash} <br />
-                            <b>Creator MSP:</b> {this.props.transaction.creator_msp_id} <br />
-                            <b>Endorser:</b> {this.props.transaction.endorser_msp_id} <br />
-                            <b>Chaincode Name:</b> {this.props.transaction.chaincodename} <br />
-                            <b>Type:</b> {this.props.transaction.type} <br />
-                            <b>Time:</b> {moment(this.props.transaction.createdt).tz(moment.tz.guess()).format("M-D-YYYY h:mm A zz")} <br />
-                            <b>Reads:</b>
-                             <ul>
-                                {this.props.transaction.read_set.map(function (item) {
+  componentWillReceiveProps(nextProps) {
+    this.setState({ loading: false });
+  }
 
-                                    return item === null ? '' :
-                                        <li><Typography variant="subheading"> {item.chaincode}</Typography>
-                                            <ul>{item.set.map(function (x) {
-                                                var block_num = '';
-                                                var tx_num = '';
-                                                if (x.version !== null) {
-                                                    block_num = x.version.block_num;
-                                                    tx_num = x.version.tx_num;
-                                                }
-                                                return x === null ? '' : <li>key:{x.key} ,version:( block:{block_num},tx:{tx_num})  </li>
-                                            })}</ul>
-                                            <br />
-                                        </li>;
-                                })}
-                            </ul>
-                            <b>Writes:</b>
+  handleClose = () => {
+    this.props.onClose();
+  };
+
+  render() {
+    const { classes } = this.props;
+    if (this.props.transaction.read_set === undefined) {
+      return (
+        <div>
+          <div>
+            <CardTitle className="dialogTitle">
+              <FontAwesome name="list-alt" className="listIcon" />Transaction
+              Details
+              <button onClick={this.handleClose} className="closeBtn">
+                <FontAwesome name="close" />
+              </button>
+            </CardTitle>
+            <div align="center">
+              <CardBody>
+                <span className="loading-wheel">
+                  {" "}
+                  <FontAwesome name="circle-o-notch" size="3x" spin />
+                </span>
+              </CardBody>
+            </div>
+          </div>
+        </div>
+      );
+    } else {
+      return (
+        <div className="dialog">
+          <Card>
+            <CardTitle className="dialogTitle">
+              <FontAwesome name="list-alt" className="listIcon" />Transaction
+              Details
+              <button onClick={this.handleClose} className="closeBtn">
+                <FontAwesome name="close" />
+              </button>
+            </CardTitle>
+            <CardBody>
+              <Table striped hover responsive className="table-striped">
+                <tbody>
+                  <tr>
+                    <th>Transaction ID:</th>
+                    <td>
+                      {this.props.transaction.txhash}
+                      <button className="copyBtn">
+                        <CopyToClipboard text={this.props.transaction.txhash}>
+                          <FontAwesome name="copy" />
+                        </CopyToClipboard>
+                      </button>
+                    </td>
+                  </tr>
+                  <tr>
+                    <th>Creator MSP:</th>
+                    <td>{this.props.transaction.creator_msp_id}</td>
+                  </tr>
+                  <tr>
+                    <th>Endoser:</th>
+                    <td>{this.props.transaction.endorser_msp_id}</td>
+                  </tr>
+                  <tr>
+                    <th>Chaincode Name:</th>
+                    <td>{this.props.transaction.chaincodename}</td>
+                  </tr>
+                  <tr>
+                    <th>Type:</th>
+                    <td>{this.props.transaction.type}</td>
+                  </tr>
+                  <tr>
+                    <th>Time:</th>
+                    <td>
+                      {moment(this.props.transaction.createdt)
+                        .tz(moment.tz.guess())
+                        .format("M-D-YYYY h:mm A zz")}
+                    </td>
+                  </tr>
+                  <tr>
+                    <th style={reads}>Reads:</th>
+                    <td>
+                      {" "}
+                      {this.props.transaction.read_set.map(function(
+                        item,
+                        index
+                      ) {
+                        return item === null ? (
+                          ""
+                        ) : (
+                          <li key={index}>
+                            <Typography variant="subheading">
+                              {" "}
+                              {item.chaincode}
+                            </Typography>
                             <ul>
-                                {this.props.transaction.write_set.map(function (item) {
-
-                                    return item === null ? '' :
-                                        <li><Typography variant="subheading"> {item.chaincode}</Typography>
-                                            <ul>{item.set.map(function (x) {
-                                                return x === null ? '' : <li>key:{x.key} ,is_delete:{x.is_delete.toString()},value:{x.value}  </li>
-                                            })}</ul>
-                                            <br />
-                                        </li>;
-                                })}
+                              {item.set.map(function(x, index) {
+                                var block_num = "";
+                                var tx_num = "";
+                                if (x.version !== null) {
+                                  block_num = x.version.block_num;
+                                  tx_num = x.version.tx_num;
+                                }
+                                return x === null ? (
+                                  ""
+                                ) : (
+                                  <li key={index}>
+                                    key:{x.key} ,version:( block:{block_num},tx:{
+                                      tx_num
+                                    }){" "}
+                                  </li>
+                                );
+                              })}
                             </ul>
-                        </DialogContentText>
-                    </DialogContent>
-                </div>
-
-            );
-        }
+                            <br />
+                          </li>
+                        );
+                      })}
+                    </td>
+                  </tr>
+                  <tr>
+                    <th style={writes}>Writes:</th>
+                    <td>
+                      {" "}
+                      {this.props.transaction.write_set.map(function(
+                        item,
+                        index
+                      ) {
+                        return item === null ? (
+                          ""
+                        ) : (
+                          <li key={index}>
+                            <Typography variant="subheading">
+                              {" "}
+                              {item.chaincode}
+                            </Typography>
+                            <ul>
+                              {item.set.map(function(x, index) {
+                                return x === null ? (
+                                  ""
+                                ) : (
+                                  <li key={index}>
+                                    key:{x.key} ,is_delete:{x.is_delete.toString()},value:{
+                                      x.value
+                                    }{" "}
+                                  </li>
+                                );
+                              })}
+                            </ul>
+                            <br />
+                          </li>
+                        );
+                      })}
+                    </td>
+                  </tr>
+                </tbody>
+              </Table>
+            </CardBody>
+          </Card>
+        </div>
+      );
     }
+  }
 }
 
-
 TransactionView.propTypes = {
-    classes: PropTypes.object.isRequired,
+  classes: PropTypes.object.isRequired
 };
 
 export default withStyles(styles)(TransactionView);
+
