@@ -12,14 +12,15 @@ import { getChannel } from './store/actions/channel/action-creators';
 import './static/css/style.css';
 import 'bootstrap/dist/css/bootstrap.css';
 import App from './components/App/App';
-import registerServiceWorker from './registerServiceWorker';
+import { unregister } from './registerServiceWorker';
+
 const store = createStore();
 store.dispatch(getChannel());
 store.dispatch(getChannelList());
 
+unregister();
 ReactDOM.render(
     <Provider store={store} >
         <App />
     </Provider>, document.getElementById('root')
 );
-registerServiceWorker();
