@@ -6,7 +6,7 @@ import { createAction } from 'redux-actions'
 import * as actionTypes from '../action-types'
 import { get } from '../../../services/request.js';
 
-export const getBlockList = (channel) => dispatch => {
+export const blockList = (channel) => dispatch => {
     get(`/api/blockAndTxList/${channel}/0`)
         .then(resp => {
             dispatch(createAction(actionTypes.BLOCK_LIST_POST)(resp))
