@@ -6,7 +6,6 @@ import { createAction } from 'redux-actions'
 import * as actionTypes from '../action-types'
 import { get } from '../../../services/request.js';
 export const getChannel = () => dispatch => {
-    // post('curChannel')
     get('/api/curChannel')
         .then(resp => {
             dispatch(createAction(actionTypes.CHANNEL)(resp))
@@ -15,3 +14,11 @@ export const getChannel = () => dispatch => {
         })
 }
 
+export const changeChannel = (channelName) => dispatch => {
+    get('/api/changeChannel/' + channelName)
+        .then(resp => {
+            dispatch(createAction(actionTypes.CHANGECHANNEL)(resp))
+        }).catch((error) => {
+            console.error(error);
+        })
+}
