@@ -6,8 +6,7 @@ import { createAction } from 'redux-actions';
 import * as actionTypes from '../action-types';
 import { get } from '../../../services/request.js';
 
-export const getHeaderCount = (channel) => dispatch => {
-    // get('/api/status/get')
+export const countHeader = (channel) => dispatch => {
     get('/api/status/' + channel)
         .then(resp => {
             dispatch(createAction(actionTypes.COUNT_HEADER_POST)(resp))
