@@ -5,8 +5,8 @@
 import { createAction } from 'redux-actions'
 import * as actionTypes from '../action-types'
 import { get } from '../../../services/request.js';
-export const getLatestBlock = (channel) => dispatch => {
-    // post('/api/status/get')
+
+export const latestBlock = (channel) => dispatch => {
     get('/api/status/'+channel)
         .then(resp => {
             dispatch(createAction(actionTypes.LATEST_BLOCK)(resp.latestBlock))
@@ -14,4 +14,3 @@ export const getLatestBlock = (channel) => dispatch => {
             console.error(error);
         })
 }
-
