@@ -6,7 +6,7 @@ import { createAction } from 'redux-actions'
 import * as actionTypes from '../action-types'
 import { get } from '../../../services/request.js';
 
-export const getTransactionList = (channel,offset) => dispatch => {
+export const transactionList = (channel) => dispatch => {
     get('/api/txList/'+channel+'/0/0/')
         .then(resp => {
             dispatch(createAction(actionTypes.TX_LIST)(resp))
